Desktop/weather/requirements.txt
requests
streamlit
pandas
google-genai